fix: StopModules comment, ccSessionID path validation, split FakeExecutor
- core.go: fix StopModules comment — "reverse dependency order" not
  "registration order" (topo sort reorders modules)
- history.go: reject ccSessionID containing path separators (/ or \)
  to prevent path traversal via crafted session IDs
- executor.go: split FakeExecutor + test helper types to
  fake_executor.go (executor.go: 176 lines, fake_executor.go: 253 lines)
